requirements.txt
crewai
uv
python-dotenv
crewai[tools]
google-generativeai
ipywidgets
docling
duckduckgo-search
langchain-community